Another update
K, new plan.
We'll have a timer in our frame class, all it does it listen if the character has moved off screen, if it does, it removes then adds the image because for the object being removed, it will have set a new location for itself and therefore seemed like it changed locations

BOOM

then we need to validate and repaint the frame. 
Essentially we'll just create instances to the frame class of all the possible objects that can be on a screen, then remove and add them as we go. The actual objects will have the action listeners to move depending on where they are on the map